+++ interoperator/pkg/controller/add_sfservicebinding.go
@@ -19,7 +19,7 @@ limitations under the License.
 package controller
 
 import (
-	"github.com/cloudfoundry-incubator/service-fabrik-broker/interoperator/pkg/controller/sfservicebinding"
+	"github.com/cloudfoundry-incubator/service-fabrik-broker/interoperator/pkg/controller/provisioners/sfservicebinding"
 )
 
 func init() {
