+++ interoperator/pkg/controller/add_sfserviceinstance.go
@@ -19,7 +19,7 @@ limitations under the License.
 package controller
 
 import (
-	"github.com/cloudfoundry-incubator/service-fabrik-broker/interoperator/pkg/controller/sfserviceinstance"
+	"github.com/cloudfoundry-incubator/service-fabrik-broker/interoperator/pkg/controller/provisioners/sfserviceinstance"
 )
 
 func init() {
